Added 3D rendered sphere ontop of video to act as rotational trackball for the sphere/dome mesh.

+++ src/application.h
@@ -100,10 +100,10 @@ public:
         int gridToMapOn;
         double radius;
         double fov;
-        int rotateX;
-        int rotateY;
-        int rotateZ;
-        int translateY;
+        double rotateX;
+        double rotateY;
+        double rotateZ;
+        double translateY;
     };
 
     SyncHelper();
